# packages/api-contracts/openapi/platform.yaml
openapi: 3.1.0
info:
  title: Super Trunfo Platform API
  version: 0.1.0
  description: Aggregated MVP API surface for service contracts.
paths:
...
  /cards/{id}:
    get:
      summary: Get card details
      tags: [Cards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "202":
          description: Planned endpoint
...
  /match/{id}:
    get:
      summary: Get match state
      tags: [Gameplay]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "202":
          description: Planned endpoint
  /match/{id}/play:
    post:
      summary: Play a round
      tags: [Gameplay]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "202":
          description: Planned endpoint
...